SIMPLE LAUNDRY SYSTEM IN PHP WITH SOURCE CODE
The Simple Laundry System In PHP is a laundry management system in towns. The project contains the admin sides and the user side. The admin is of two types: the main admin and the sub-admin(staff). The admin can view and manage the users, sub-admin, and manage the laundry services. While the sub-admin can only manage the laundry services.

The users are clients who want to receive laundry services. A user has to login first to check services. If he/she has not registered, they can register first and then login. This project is a very simple project that solves laundry issues.

About System
This Simple Laundry System is in PHP, CSS, JavaScript, and bootstrap. Talking about the features of this system, it just contains both the admin section and the user section. The user can log in to check the status of their services, while the admins can view all users, their services to be performed and manage the system. The design of this project is pretty simple so that the user won’t find any difficulties while working on it.

How To Run The Project?
To run this project, you must have installed a virtual server i.e XAMPP on your PC (for Windows). This Simple Laundry System in PHP with source code is free to download, Use for educational purposes only!

After Starting Apache and MySQL in XAMPP, follow the following steps.

1st Step: Extract file
2nd Step: Copy the main project folder
3rd Step: Paste in xampp/htdocs/

4th Step: Open a browser and go to URL “http://localhost/phpmyadmin/”
5th Step: Then, click on the databases tab
6th Step: Create a database naming “bbn” and then click on the import tab
7th Step: Click on browse file and select “bbn.sql” file which is inside the “Laundry_system” folder
8th Step: Click on go.

After Creating Database,

9th Step: Open a browser and go to URL “http://localhost/Laundry_system/welcome.php”